+++ src/chrome/local-runtime.ts
@@ -31,6 +31,7 @@ import {
   mousedown,
   mouseup,
   focus,
+  clear
 } from '../util'
 
 export default class LocalRuntime {
@@ -94,6 +95,8 @@ export default class LocalRuntime {
         return this.mousup(command.selector)
       case 'focus':
         return this.focus(command.selector)
+      case 'clear':
+        return this.clear(command.selector)
       default:
         throw new Error(`No such command: ${JSON.stringify(command)}`)
     }
@@ -357,6 +360,22 @@ export default class LocalRuntime {
 
       return filePath
     }
+  }
+
+  async clear(selector: string): Promise<void> {
+    if (selector) {
+      if (this.chromelessOptions.implicitWait) {
+        this.log(`clear(): Waiting for ${selector}`)
+        await waitForNode(this.client, selector, this.chromelessOptions.waitTimeout)
+      }
+
+      const exists = await nodeExists(this.client, selector)
+      if (!exists) {
+        throw new Error(`clear(): Node not found for selector: ${selector}`)
+      }
+    }
+    await clear(this.client, selector)
+    this.log(`${selector} cleared`)
   }
 
   private log(msg: string): void {